Fix: Align HTTP status code for `insufficient_scope` with RFC6750
As I noted in [chat], the Micropub spec doesn't align with the OAuth2
spec regarding the `insufficient_scope` error message. We should align
with [RFC6750] to ensure that implementers can rely on out-of-the-box
OAuth2 implementations.

Additionally, move the element up, to group it with the other 403 error
code, for readability.

[chat]: https://chat.indieweb.org/dev/2020-04-11/1586648523426500
[RFC6750]: https://tools.ietf.org/html/rfc6750#section-3.1